Workspace refactor
* workspace refactor

* cr fix

* listen to isfetching

* unused import

* warn if data is stale

* refactor

* refactor

* rename

* refactor workspace api

* extended DataApi

* refactor

* cr fix

* cr fix

* remove unused styling

* Swap places between filter and refresh icon
