+++ src/Core/Client/Types/App.ts
@@ -5,7 +5,7 @@ import { WorkspaceOptions } from '../Service/ClientBuilder';
 import { AppConfig } from './AppConfig';
 import { AppManifest } from './AppManifest';
 
-type AppType = 'DataViewer' | 'PageView' | 'CustomApp' | 'PowerBI';
+type AppType = 'Workspace' | 'PageView' | 'CustomApp' | 'PowerBI';
 
 export type CustomClientApi = Omit<ClientApi, 'createWorkSpace' | 'createPageViewer'>;
