+++ src/Core/WorkSpace/src/Components/DataViewerHeader/Header.tsx
@@ -1,6 +1,10 @@
 import { useFactory } from '@equinor/DataFactory';
 import { Tabs } from '@equinor/eds-core-react';
+import { tokens } from '@equinor/eds-tokens';
 import { useFilteredData } from '@equinor/filter';
+import { DateTime } from 'luxon';
+import { useEffect, useState } from 'react';
+import { ClickableIcon } from '../../../../../components/Icon/ClickableIcon';
 import Icon from '../../../../../components/Icon/Icon';
 import { StatusBar } from '../../../../../packages/StatusBar';
 import { useDataContext } from '../../Context/DataProvider';
@@ -29,9 +33,19 @@ export const CompletionViewHeader = ({
     handleFilter,
     activeFilter,
 }: CompletionViewHeaderProps): JSX.Element => {
-    const { statusFunc, key } = useDataContext();
+    const { statusFunc, key, dataApi } = useDataContext();
     const { factory, setSelected } = useFactory(key);
     const { data } = useFilteredData();
+    const [timestamp, setTimestamp] = useState<string | null>(makeTimestamp(dataApi.dataUpdatedAt));
+
+    function makeTimestamp(timeInMs: number): string | null {
+        return DateTime.fromMillis(timeInMs).toRelative({ unit: 'minutes' });
+    }
+
+    useEffect(() => {
+        setTimestamp(makeTimestamp(dataApi.dataUpdatedAt));
+        setInterval(() => setTimestamp(makeTimestamp(dataApi.dataUpdatedAt)), 1000 * 60);
+    }, [dataApi.dataUpdatedAt]);
 
     return (
         <HeaderWrapper>
@@ -64,6 +78,22 @@ export const CompletionViewHeader = ({
                     })}
                 </List>
                 <Divider />
+                <TabButton
+                    color={
+                        dataApi.isStale
+                            ? tokens.colors.infographic.primary__energy_red_100.hex
+                            : 'grey'
+                    }
+                    aria-selected={false}
+                    title={
+                        dataApi.isStale
+                            ? 'This data is over 1 hour old and might be outdated'
+                            : `Updated: ${timestamp}`
+                    }
+                    onClick={() => dataApi.refetch()}
+                >
+                    <ClickableIcon size={32} name="refresh" />
+                </TabButton>
                 <TabButton onClick={handleFilter} aria-selected={activeFilter} title="Filter">
                     <Icon name={'filter_alt'} />
                 </TabButton>
