+++ src/Core/WorkSpace/src/Components/DataViewerHeader/HeaderStyles.ts
@@ -18,6 +18,7 @@ export const LeftSection = styled.div`
     width: -webkit-fill-available;
     border-bottom: 2px solid ${tokens.colors.ui.background__medium.rgba};
 `;
+
 export const RightSection = styled.div`
     display: flex;
     flex-direction: row;
