+++ src/Core/WorkSpace/src/WorkSpaceApi/useWorkSpace.ts
@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
 import { useAtom } from '@dbeining/react-atom';
 import { useMemo } from 'react';
-import { useWorkSpaceKey } from '../Components/DefaultView/Hooks/useDataViewerKey';
-import { getWorkSpaceContext, WorkSpaceConfig } from './State';
+import { useWorkSpaceKey } from '../Components/DefaultView/Hooks/useWorkspaceKey';
+import { getWorkSpaceContext, WorkSpaceConfig } from './workspaceState';
 
 export function useWorkSpace<T>(): WorkSpaceConfig<T> {
     const key = useWorkSpaceKey();
@@ -12,9 +12,9 @@ export function useWorkSpace<T>(): WorkSpaceConfig<T> {
     if (currentView) {
         return currentView as WorkSpaceConfig<T>;
     } else {
-        console.warn(`No DataView registered on path/key:  ${key}`);
+        console.warn(`No workspace registered on path/key:  ${key}`);
         return {
-            name: `Unknown DataView ${key}`,
+            name: `Unknown Workspace ${key}`,
             objectIdentifier: 'Unknown objectIdentifier',
         };
     }
