+++ src/apps/ScopeChangeRequest/WorkspaceConfig/Table/getLastSigned.ts
@@ -1,11 +1,11 @@
 import { DateTime } from 'luxon';
-import { ScopeChangeRequest } from '../Types/scopeChangeRequest';
+import { ScopeChangeRequest } from '../../Types/scopeChangeRequest';
 
 export function getLastSigned(request: ScopeChangeRequest): DateTime | null {
     if (!request || request.state === 'Draft') return null;
 
     const dateArray: DateTime[] = [];
-
+    if (!request.workflowSteps) return null;
     request.workflowSteps.forEach((step) =>
         step.criterias.forEach((criteria) => {
             if (criteria.signedAtUtc) {
