+++ src/components/Icon/ClickableIcon.tsx
@@ -5,6 +5,7 @@ export interface ClickableIconProps {
     name: string;
     onClick?: React.MouseEventHandler<SVGSVGElement>;
     color?: string | undefined;
+    size?: 16 | 24 | 32 | 40 | 48;
 }
 
 /**
@@ -17,6 +18,15 @@ export const ClickableIcon = ({
     name,
     onClick,
     color = `${tokens.colors.interactive.primary__resting.hex}`,
+    size = 24,
 }: ClickableIconProps): JSX.Element => {
-    return <Icon name={name} onClick={onClick} style={{ cursor: 'pointer' }} color={color} />;
+    return (
+        <Icon
+            size={size}
+            name={name}
+            onClick={onClick}
+            style={{ cursor: 'pointer' }}
+            color={color}
+        />
+    );
 };
